fix(watch): bar the rendition switch on the lower of live and the outgoing playhead
Measuring the incoming rendition against the live playhead alone regressed
two cases. The sync reference only ever moves down, so when delivery runs
late both playheads sit behind live and neither can reach it, stalling the
switch exactly when ABR wants to down-switch. And the sync buffer already
carries the incoming rendition's jitter (the source publishes it as soon as
the target changes, and now() subtracts the buffer), so allowing that jitter
again counted it twice: a 2s segmented rendition passed on its very first
rendered frame and stepped the picture 2s backwards.

Take the lower of the two bars instead. The outgoing playhead alone was the
original defect: playheads never rewind, so switching to a coarser rendition
grows the buffer and strands that playhead ahead of live, freezing the
picture until wall-clock time makes up the difference. Neither bar works
alone; the lower of the two covers both.

Per-rendition jitter keeps its role through the sync buffer, which is what
places live for the incoming rendition, so the decision no longer needs it
directly. Renames the module to playhead.ts to match what it now holds.
