+++ js/watch/src/video/decoder.ts
@@ -7,7 +7,7 @@ import { Effect, type Getter, getter, type Inputs, type Readonlys, readonlys, Si
 import { base64ToBytes } from "../base64";
 
 import type { Sync } from "../sync";
-import { caughtUp, renditionJitter } from "./jitter";
+import { caughtUp } from "./playhead";
 import { rotateVideoDimensions } from "./presentation";
 import type { Source } from "./source";
 
@@ -138,16 +138,10 @@ export class Decoder {
 				const pendingTimestamp = effect.get(pending.timestamp);
 				if (!pendingTimestamp) return;
 
-				// Switch once the new rendition has caught up to live, judged against its own
-				// jitter. Live comes from the shared clock rather than the outgoing track's
-				// playhead: that playhead trails live by its own jitter, so comparing the two
-				// directly reads a coarse rendition as caught up early (whenever the outgoing
-				// one is between groups) and a fine one as never caught up. Before the clock
-				// has an anchor there is no live edge, so fall back to the outgoing playhead.
-				const live = this.sync.now() ?? effect.get(current.timestamp);
-				if (live !== undefined && !caughtUp({ playhead: pendingTimestamp, live, jitter: pending.jitter })) {
-					return;
-				}
+				// Hold off until the new rendition has caught up to the picture it's replacing.
+				const active = effect.get(current.timestamp);
+				const live = this.sync.now();
+				if (!caughtUp({ playhead: pendingTimestamp, active, live })) return;
 			}
 
 			// Upgrade the pending track to active.
@@ -245,9 +239,6 @@ class DecoderTrack {
 	config: RequiredDecoderConfig;
 	stats: Signal<Stats | undefined>;
 
-	// How far behind live this rendition's playhead can sit while still being at its live edge.
-	jitter: Time.Milli;
-
 	timestamp = new Signal<Time.Milli | undefined>(undefined);
 	frame = new Signal<VideoFrame | undefined>(undefined);
 
@@ -272,7 +263,6 @@ class DecoderTrack {
 		this.track = props.track;
 		this.config = requiredConfig;
 		this.stats = props.stats;
-		this.jitter = renditionJitter(props.config) ?? Time.Milli.zero;
 
 		this.#signals.run(this.#run.bind(this));
 	}
